requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
apscheduler>=3.10.1
cryptography>=41.0.3
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
python-dateutil>=2.8.2
SQLAlchemy>=2.0.0
alembic>=1.12.1
aiosqlite>=0.19.0
asyncpg>=0.28.0  # For PostgreSQL
SQLAlchemy[asyncio]>=2.0.0
python-telegram-bot>=20.6
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.0.0
uvicorn>=0.24.0
